requirements.txt
paddlepaddle
paddleocr
uiautomator2
uiautodev
numpy
pyyaml
setuptools
opencv-python